+++ rana_qgis_plugin/loader.py
@@ -52,6 +52,7 @@ from rana_qgis_plugin.utils_api import (
     create_folder,
     delete_tenant_project_directory,
     delete_tenant_project_file,
+    get_process_id_for_tag,
     get_tenant_details,
     get_tenant_file_descriptor,
     get_tenant_file_descriptor_view,
@@ -79,6 +80,7 @@ from rana_qgis_plugin.workers import (
     FileUploadWorker,
     LizardResultDownloadWorker,
     ProjectJobMonitorWorker,
+    RasterStyleWorker,
     VectorStyleWorker,
 )
 
@@ -121,9 +123,6 @@ class PyQtSlotExceptionHandlingMeta(QObjectMeta):
             try:
                 return method(self, *args, **kwargs)
             except Exception as e:
-                # Get the error message
-                error_message = f"{e.__class__.__name__}: {str(e)}"
-
                 # Get the full traceback as a separate string
                 tb = traceback.format_exc()
 
@@ -146,6 +145,8 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
     new_file_upload_finished = pyqtSignal(str)
     vector_style_finished = pyqtSignal()
     vector_style_failed = pyqtSignal(str)
+    raster_style_finished = pyqtSignal()
+    raster_style_failed = pyqtSignal(str)
     loading_cancelled = pyqtSignal()
     download_results_cancelled = pyqtSignal()
     schematisation_upload_cancelled = pyqtSignal()
@@ -166,12 +167,14 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
     model_deleted = pyqtSignal()
     project_jobs_added = pyqtSignal(list)
     project_job_updated = pyqtSignal(dict)
+    file_opened = pyqtSignal(dict)
 
     def __init__(self, communication, parent):
         super().__init__(parent)
         self.file_download_worker: QThread = None
         self.file_upload_worker: QThread = None
         self.vector_style_worker: QThread = None
+        self.raster_style_worker: QThread = None
         self.new_file_upload_worker: QThread = None
         self.project_job_monitor: QThread = None
         self.communication = communication
@@ -207,6 +210,7 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
                         "wms",
                     )
                     QgsProject.instance().addMapLayer(rlayer)
+                self.file_opened.emit(file)
                 return True
 
         return False
@@ -258,6 +262,7 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
             hcc_working_dir(),
             get_threedi_api(),
         )
+        self.file_opened.emit(file)
         self.file_download_finished.emit(None)
 
     def on_file_download_finished(
@@ -319,6 +324,7 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
                 get_tenant_file_descriptor(file["descriptor_id"]),
                 schematisation,
             )
+        self.file_opened.emit(file)
         self.file_download_finished.emit(local_file_path)
 
     def on_file_download_failed(self, error: str):
@@ -617,12 +623,6 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
             if simulation_wizard.exec() == QDialog.Rejected:
                 self.simulation_cancelled.emit()
 
-    def get_process_id_for_tag(self, tag: str) -> Optional[str]:
-        processes = get_tenant_processes(self.communication)
-        for process in processes:
-            if tag in process["tags"]:
-                return process["id"]
-
     def start_model_tracker_process(
         self,
         project,
@@ -630,7 +630,7 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
         revision_id: int,
         inherit_from_previous_revision: bool = True,
     ):
-        track_process = self.get_process_id_for_tag("model_tracker")
+        track_process = get_process_id_for_tag(self.communication, "model_tracker")
         if track_process is None:
             self.communication.log_err("No model tracker available")
             return
@@ -649,7 +649,7 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
 
     def start_simulation_tracker_process(self, project, file, simulations):
         # Find the simulation tracker processes
-        track_process = self.get_process_id_for_tag("simulation_tracker")
+        track_process = get_process_id_for_tag(self.communication, "simulation_tracker")
 
         if track_process is None:
             self.communication.log_err("No simulation tracker available")
@@ -972,7 +972,7 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
         sender = self.sender()
         assert isinstance(sender, QThread)
         sender.wait()
-
+        sender.deleteLater()
         self.file_upload_finished.emit()
 
     def on_new_file_upload_finished(self, online_path: str, project):
@@ -1010,6 +1010,21 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
         self.vector_style_worker.warning.connect(self.communication.show_warn)
         self.vector_style_worker.start()
 
+    @pyqtSlot(dict, dict)
+    def save_raster_style(self, project, file):
+        """Start the worker for saving raster styling files"""
+        self.communication.progress_bar(
+            "Generating and saving raster styling files...", clear_msg_bar=True
+        )
+        self.raster_style_worker = RasterStyleWorker(
+            project,
+            file,
+        )
+        self.raster_style_worker.finished.connect(self.on_raster_style_finished)
+        self.raster_style_worker.failed.connect(self.on_raster_style_failed)
+        self.raster_style_worker.warning.connect(self.communication.show_warn)
+        self.raster_style_worker.start()
+
     def on_vector_style_finished(self, msg: str):
         self.communication.clear_message_bar()
         self.communication.show_info(msg)
@@ -1019,6 +1034,16 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
         self.communication.clear_message_bar()
         self.communication.show_error(msg)
         self.vector_style_failed.emit(msg)
+
+    def on_raster_style_finished(self, msg: str):
+        self.communication.clear_message_bar()
+        self.communication.show_info(msg)
+        self.raster_style_finished.emit()
+
+    def on_raster_style_failed(self, msg: str):
+        self.communication.clear_message_bar()
+        self.communication.show_error(msg)
+        self.raster_style_failed.emit(msg)
 
     @pyqtSlot(dict, dict)
     def import_schematisation_to_rana(self, project, selected_file):
@@ -1080,7 +1105,9 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
             # Search GeoPackage database tables for attributes with file paths.
             paths = new_schematisation_wizard.get_paths_from_geopackage(db_path)
 
-            self.save_initial_revision(new_schematisation, local_schematisation, paths)
+            self.save_initial_revision(
+                project, new_schematisation, local_schematisation, paths
+            )
 
         if rana_path:
             file_path = rana_path + new_schematisation.name
@@ -1249,8 +1276,10 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
             clear_msg_bar=True,
         )
 
-    @pyqtSlot(dict, dict, dict, dict)
-    def save_initial_revision(self, schematisation, local_schematisation, raster_paths):
+    @pyqtSlot(dict, dict, dict, dict, dict)
+    def save_initial_revision(
+        self, project, schematisation, local_schematisation, raster_paths
+    ):
         raster_dir = local_schematisation.wip_revision.raster_dir
 
         selected_files = {
@@ -1295,7 +1324,7 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
             "selected_files": selected_files,
             "commit_message": "Initial commit",
             "create_revision": True,
-            "make_3di_model": False,
+            "make_3di_model": True,
             "cb_inherit_templates": False,
         }
 
@@ -1306,6 +1335,14 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
             local_schematisation,
             upload_template,
         )
+        upload_worker.signals.create_model_requested.connect(
+            lambda revision_id: self.start_model_tracker_process(
+                project,
+                schematisation.to_dict(),
+                revision_id,
+            )
+        )
+
         upload_worker.signals.thread_finished.connect(
             self.schematisation_upload_finished
         )
@@ -1330,3 +1367,20 @@ class Loader(QObject, metaclass=PyQtSlotExceptionHandlingMeta):
         self.project_job_monitor.job_updated.connect(self.project_job_updated)
         self.project_job_monitor.failed.connect(self.communication.show_warn)
         self.project_job_monitor.start()
+
+    @pyqtSlot(int)
+    def cancel_simulation(self, simulation_pk):
+        confirm_cancel = QMessageBox.warning(
+            None,
+            "Cancel Simulation",
+            "Are you sure you want to cancel the simulation?",
+            QMessageBox.StandardButton.Yes,
+            QMessageBox.StandardButton.No,
+        )
+        if confirm_cancel == QMessageBox.StandardButton.Yes:
+            tc = ThreediCalls(get_threedi_api())
+            tc.fetch_simulation_status(simulation_pk)
+            try:
+                tc.create_simulation_action(simulation_pk, name="shutdown")
+            except ApiException as e:
+                self.communication.show_error(f"Could not cancel simulation")