+++ rana_qgis_plugin/rana_qgis_plugin.py
@@ -255,6 +255,9 @@ class RanaQgisPlugin:
             self.loader.project_job_updated.connect(
                 self.rana_browser.project_job_updated.emit
             )
+            self.rana_browser.processes_browser.cancel_simulation.connect(
+                self.loader.cancel_simulation
+            )
             self.rana_browser.open_wms_selected.connect(self.loader.open_wms)
             self.rana_browser.open_in_qgis_selected.connect(self.rana_browser.disable)
             self.rana_browser.open_in_qgis_selected.connect(self.loader.open_in_qgis)
@@ -263,6 +266,9 @@ class RanaQgisPlugin:
             )
             self.rana_browser.save_vector_styling_selected.connect(
                 self.loader.save_vector_style
+            )
+            self.rana_browser.save_raster_styling_selected.connect(
+                self.loader.save_raster_style
             )
             self.rana_browser.upload_new_file_selected.connect(
                 self.rana_browser.disable
@@ -297,6 +303,9 @@ class RanaQgisPlugin:
             )
             self.rana_browser.upload_file_selected.connect(self.rana_browser.disable)
             self.rana_browser.save_vector_styling_selected.connect(
+                self.rana_browser.disable
+            )
+            self.rana_browser.save_raster_styling_selected.connect(
                 self.rana_browser.disable
             )
             self.loader.download_results_cancelled.connect(self.rana_browser.enable)
@@ -340,6 +349,8 @@ class RanaQgisPlugin:
                 self.rana_browser.show_processes_overview
             )
             self.loader.file_download_finished.connect(self.rana_browser.enable)
+            self.loader.file_opened.connect(self.rana_browser.view_file_after_open)
+            self.loader.file_download_finished.connect(self.rana_browser.refresh)
             self.loader.file_download_failed.connect(self.rana_browser.enable)
             self.loader.file_upload_finished.connect(self.rana_browser.enable)
             self.loader.file_upload_finished.connect(self.rana_browser.refresh)
@@ -350,6 +361,9 @@ class RanaQgisPlugin:
             self.loader.vector_style_finished.connect(self.rana_browser.enable)
             self.loader.vector_style_finished.connect(self.rana_browser.refresh)
             self.loader.vector_style_failed.connect(self.rana_browser.enable)
+            self.loader.raster_style_finished.connect(self.rana_browser.enable)
+            self.loader.raster_style_finished.connect(self.rana_browser.refresh)
+            self.loader.raster_style_failed.connect(self.rana_browser.enable)
             self.loader.simulation_started.connect(self.rana_browser.enable)
             self.loader.simulation_cancelled.connect(self.rana_browser.enable)
             self.loader.simulation_started_failed.connect(self.rana_browser.enable)