+++ rana_qgis_plugin/utils_api.py
@@ -415,6 +415,21 @@ def get_vector_style_upload_urls(descriptor_id: str):
         return None
 
 
+def get_raster_style_upload_urls(descriptor_id: str, files):
+    authcfg_id = get_authcfg_id()
+    tenant = get_tenant_id()
+    url = f"{api_url()}/tenants/{tenant}/file-descriptors/{descriptor_id}/raster-style"
+
+    network_manager = NetworkManager(url, authcfg_id)
+    status = network_manager.put_multipart(files=files)
+
+    if status:
+        response = network_manager.content
+        return response
+    else:
+        return None
+
+
 def get_style_file(source_type: str, descriptor_id: str, file_name: str):
     authcfg_id = get_authcfg_id()
     tenant = get_tenant_id()
@@ -568,3 +583,10 @@ def get_project_jobs(project_id: str):
         return network_manager.content
     else:
         return None
+
+
+def get_process_id_for_tag(communication: UICommunication, tag: str) -> Optional[str]:
+    processes = get_tenant_processes(communication)
+    for process in processes:
+        if tag in process["tags"]:
+            return process["id"]